+++ app/src/js/utils/table-config/pdr-progress.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { get } from 'object-path';
-import { tally, bool, fromNow, nullValue } from '../format';
+import { tally, bool, nullValue, fromNowWithTooltip } from '../format';
 import ErrorReport from '../../components/Errors/report';
 import { getPersistentQueryParams } from '../url-helper';
 
@@ -83,7 +83,8 @@ export const tableColumns = [
   },
   {
     Header: 'Discovered',
-    accessor: (row) => fromNow(row.timestamp),
+    accessor: 'timestamp',
+    Cell: ({ cell: { value } }) => fromNowWithTooltip(value),
     id: 'timestamp'
   }
 ];