+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- **CUMULUS-2091**
+  - Add Tooltip component
+  - Add blue tooltip for timestamp values in the table
+
 - **CUMULUS-2200**
   - Add Granule List page. Use Granule Inventory reports for generating granule CSVs